refactor: split continuum json type carriers
