+++ src/infrastructure/adapters/GeneratedLegInventory.ts
@@ -1,0 +1,4 @@
+export type GeneratedLegInventory = {
+  readonly names: readonly string[];
+  readonly files: readonly string[];
+};
