+++ src/infrastructure/adapters/JsonObject.ts
@@ -1,0 +1,1 @@
+export type JsonObject = Readonly<Record<string, unknown>>;
