+++ src/infrastructure/adapters/WesleyIntegrityFields.ts
@@ -1,0 +1,7 @@
+export type WesleyIntegrityFields = {
+  readonly integrityStatus: string;
+  readonly integrityScope: string;
+  readonly hashAlgorithm: string;
+  readonly signatureAlgorithm: string;
+  readonly signatureKeyId: string;
+};
